refactor: rename `_vt_abc` to `_abc`
