+++ src/gwkokab/vts/_neuralvt.py
@@ -19,8 +19,8 @@ import jax
 from jax import numpy as jnp
 from jaxtyping import Array
 
+from ._abc import VolumeTimeSensitivityInterface
 from ._utils import load_model
-from ._vt_abc import VolumeTimeSensitivityInterface
 
 
 class NeuralNetVolumeTimeSensitivity(VolumeTimeSensitivityInterface):
